jobs controllers setup

+++ server/controllers/jobs.ts
@@ -1,50 +1,68 @@
 import { Request, Response, NextFunction } from 'express'
+import StatusCodes from 'http-status-codes'
+import { BadRequestError, NotFoundError } from '../errors/index.js'
 import Job from '../models/job.js'
 
-export const getJobs = async (req: Request, res: Response): Promise<any> => {
-  const jobs = await Job.find({})
-  res.status(200).json({ jobs })
+export const getJobs = async (req: any, res: Response): Promise<any> => {
+  const jobs = await Job.find({ createdBy: req.user.userId }).sort('createdAt')
+  res.status(StatusCodes.OK).json({ jobs, count: jobs.length })
 }
 
-export const createJob = async (req: Request, res: Response): Promise<any> => {
+export const getJob = async (req: any, res: Response): Promise<any> => {
+  const {
+    user: { userId },
+    params: { id: jobId },
+  } = req
+
+  const job = await Job.findOne({ _id: jobId, createdBy: userId })
+
+  if (!job) {
+    throw new NotFoundError(`No job with id ${jobId}`)
+  }
+
+  res.status(StatusCodes.OK).json({ job })
+}
+
+export const createJob = async (req: any, res: Response): Promise<any> => {
+  req.body.createdBy = req.user.userId
   const job = await Job.create(req.body)
-  res.status(201).json({ job })
+  res.status(StatusCodes.CREATED).json({ job })
 }
 
-export const getJob = async (req: Request, res: Response): Promise<any> => {
-  const jobId = req.params.id
-  const job = await Job.findOne({ _id: jobId })
+export const updateJob = async (req: any, res: Response): Promise<any> => {
+  const {
+    user: { userId },
+    params: { id: jobId },
+    body: { company, position },
+  } = req
 
-  if (!job) {
-    return res.status(404).json({ error: `Job with id ${jobId} not found` })
+  if (company === '' || position === '') {
+    throw new BadRequestError('Company or Position fields cannot be empty')
   }
 
-  res.status(200).json({ job })
-}
-
-export const updateJob = async (req: Request, res: Response): Promise<any> => {
-  const jobId = req.params.id
-  const job = await Job.findOne({ _id: jobId })
-
-  if (!job) {
-    return res.status(404).json({ error: `Job with id ${jobId} not found` })
-  }
-
-  const updatedJob = await Job.findOneAndUpdate({ _id: jobId }, req.body, {
+  const job = await Job.findOneAndUpdate({ _id: jobId, createdBy: userId }, req.body, {
     new: true,
     runValidators: true,
   })
 
-  res.status(200).json({ job: updatedJob })
-}
-
-export const deleteJob = async (req: Request, res: Response): Promise<any> => {
-  const jobId = req.params.id
-  const job = await Job.findOneAndDelete({ _id: jobId })
-
   if (!job) {
-    return res.status(404).json({ error: `Job with id ${jobId} not found` })
+    throw new NotFoundError(`No job with id ${jobId}`)
   }
 
-  res.status(200).json({ job })
+  res.status(StatusCodes.OK).json({ job })
+}
+
+export const deleteJob = async (req: any, res: Response): Promise<any> => {
+  const {
+    user: { userId },
+    params: { id: jobId },
+  } = req
+
+  const job = await Job.findOneAndDelete({ _id: jobId, createdBy: userId })
+
+  if (!job) {
+    throw new NotFoundError(`No job with id ${jobId}`)
+  }
+
+  res.status(StatusCodes.OK).json({ job })
 }
